Improves the instrumentation test and remove pointless files
