--- android/src/main/java/com/example/test/MainActivity.java
@@ -1,15 +1,0 @@
-package com.example.test;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-public class MainActivity extends Activity
-{
-    /** Called when the activity is first created. */
-    @Override
-    public void onCreate(Bundle savedInstanceState)
-    {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.main);
-    }
-}
